Move the Faraday UART library into the faraday module
This moves Proxy/faraday_uart_stack to faraday/uart and updates proxy.py
to import it from there.
