+++ faraday/uart/layer_2_protocol.py
